+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/preferences/LangSDKConfigBlock.java
@@ -14,12 +14,12 @@ import org.eclipse.jface.layout.GridLayoutFactory;
 import org.eclipse.swt.widgets.Composite;
 
 import melnorme.util.swt.SWTFactoryUtil;
-import melnorme.util.swt.components.AbstractComponentExt;
+import melnorme.util.swt.components.AbstractComponent;
 import melnorme.util.swt.components.AbstractFieldComponent;
 import melnorme.util.swt.components.fields.ButtonTextField;
 import melnorme.util.swt.components.fields.DirectoryTextField;
 
-public class LangSDKConfigBlock extends AbstractComponentExt {
+public class LangSDKConfigBlock extends AbstractComponent {
 	
 	protected final LangSDKConfigBlock.LanguageSDKLocationGroup sdkLocationGroup = createSDKLocationGroup();
 	
@@ -45,7 +45,7 @@ public class LangSDKConfigBlock extends AbstractComponentExt {
 	}
 	
 
-	public static class LanguageSDKLocationGroup extends AbstractComponentExt {
+	public static class LanguageSDKLocationGroup extends AbstractComponent {
 		
 		protected final ButtonTextField sdkLocation = createSdkLocationField();