+++ plugin_ide.ui/src-lang/melnorme/util/swt/components/AbstractFieldComponent.java
@@ -10,16 +10,10 @@
  *******************************************************************************/
 package melnorme.util.swt.components;
 
-import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.ModifyEvent;
 import org.eclipse.swt.events.ModifyListener;
-import org.eclipse.swt.events.SelectionAdapter;
-import org.eclipse.swt.events.SelectionEvent;
-import org.eclipse.swt.widgets.Button;
-import org.eclipse.swt.widgets.Combo;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
-import org.eclipse.swt.widgets.Spinner;
 import org.eclipse.swt.widgets.Text;
 
 import melnorme.util.swt.SWTUtil;
@@ -31,7 +25,7 @@ import melnorme.utilbox.fields.IFieldValueListener;
  * Field component with a field value that can be manipulated (get/set) even if the 
  * componented is not created.
  */
-public abstract class AbstractFieldComponent<VALUE> extends AbstractComponentExt 
+public abstract class AbstractFieldComponent<VALUE> extends AbstractComponent 
 	implements IDomainField<VALUE>, IWidgetComponent {
 	
 	private final DomainField<VALUE> domainField = new DomainField<>();
@@ -67,7 +61,7 @@ public abstract class AbstractFieldComponent<VALUE> extends AbstractComponentExt
 	}
 	
 	/** Update the field value from a control modification. */
-	protected void setFieldValueFromControl(VALUE newValue) {
+	public void setFieldValueFromControl(VALUE newValue) {
 		settingValueFromControl = true;
 		try {
 			doSetFieldValue(newValue);
@@ -116,7 +110,7 @@ public abstract class AbstractFieldComponent<VALUE> extends AbstractComponentExt
 	
 	/* ----------------- create helpers ----------------- */
 	
-	protected static Text createFieldText(final AbstractFieldComponent<String> field, Composite parent, int style) {
+	public static Text createFieldText(AbstractFieldComponent<String> field, Composite parent, int style) {
 		final Text text = new Text(parent, style);
 		text.addModifyListener(new ModifyListener() {
 			@Override
@@ -125,42 +119,6 @@ public abstract class AbstractFieldComponent<VALUE> extends AbstractComponentExt
 			}
 		});
 		return text;
-	}
-	
-	protected static Button createFieldCheckbox(final AbstractFieldComponent<Boolean> field, Composite parent, 
-			int style) {
-		final Button checkBox = new Button(parent, SWT.CHECK | style);
-		checkBox.addSelectionListener(new SelectionAdapter() {
-			@Override
-			public void widgetSelected(SelectionEvent e) {
-				field.setFieldValueFromControl(checkBox.getSelection());
-			}
-		});
-		return checkBox;
-	}
-	
-	protected static Spinner createFieldSpinner(final AbstractFieldComponent<Integer> field, Composite parent, 
-			int style) {
-		final Spinner spinner = new Spinner(parent, style);
-		spinner.addModifyListener(new ModifyListener() {
-			@Override
-			public void modifyText(ModifyEvent e) {
-				field.setFieldValueFromControl(spinner.getSelection());
-			}
-		});
-		return spinner;
-	}
-	
-	protected static Combo createFieldCombo(final AbstractFieldComponent<Integer> field, Composite topControl, int style) {
-		final Combo combo = new Combo(topControl, style);
-		combo.addSelectionListener(new SelectionAdapter() {
-			@Override
-			public void widgetSelected(SelectionEvent e) {
-				int selectionIndex = combo.getSelectionIndex();
-				field.setFieldValueFromControl(selectionIndex);
-			}
-		});
-		return combo;
 	}
 	
 }