+++ documentation/ChangeLog.md
@@ -1,6 +1,8 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Fixed: Arguments field in launch configuration is not multi-line.
+ 
 
 ### 0.11.2
   ▶ Recommended/tested CDT version is now 8.7